+++ src/Overlay/Loader/Loader.stories.tsx
@@ -1,0 +1,16 @@
+import type { Meta, StoryObj } from '@storybook/react'
+import { Loader } from './Loader'
+
+const meta: Meta<typeof Loader> = {
+  component: Loader,
+}
+
+export default meta
+
+type Story = StoryObj<typeof Loader>
+
+export const Default: Story = {
+  args: {
+    message: 'Loading...',
+  },
+}
